requirements.txt
numpy==1.21.5
pandas<1.4.4
pandas_ta==0.3.14b0
plotly==5.9.0
requests==2.28.1
streamlit==1.16.0
python-dotenv ==1.0.0